+++ Winnie/Core/Orc.cs
@@ -45,7 +45,7 @@ namespace Core
             else { return 1; }
         }
 
-        public float GetVictoryPoints(TileType tileType)
+        public int GetVictoryPoints(TileType tileType)
         {
             if (tileType is PlainTileType || tileType is ForestTileType) { return 1; }
             else if (tileType is MountainTileType) { return 2; }
